@@ -1,26 +1,29 @@
-var React = require('react/addons');
-var PureRenderMixin = React.addons.PureRenderMixin;
+import React from 'react';
 
-var Item = React.createClass({
+export default class Item extends React.Component {
 
-	mixins: [ PureRenderMixin ],
+	constructor(props) {
+		super(props);
+		this.props = {};
+		this.state = {};
+	}
 
-	_itemWidth: function() {
+	_itemWidth() {
 		return this.props.dimensions.gridWidth / this.props.dimensions.itemsPerRow;
-	},
+	}
 
-	_itemLeft: function() {
+	_itemLeft() {
 		var column = this.props.index % this.props.dimensions.itemsPerRow;
 		return column * (this.props.dimensions.gridWidth / this.props.dimensions.itemsPerRow);
-	},
+	}
 
-	_itemTop: function() {
+	_itemTop() {
 		return Math.floor(this.props.index / this.props.dimensions.itemsPerRow) * this.props.dimensions.height;
-	},
+	}
 
-	render: function() {
+	render() {
 
-		var _style = {
+		const _style = {
 			width: this._itemWidth() - this.props.padding,
 			height: this.props.dimensions.height - this.props.padding,
 			left: this._itemLeft(),
@@ -28,15 +31,16 @@ var Item = React.createClass({
 			position: "absolute"
 		};
 
-		var ItemRenderer = this.props.ItemRenderer || "div";
-
 		var props = {
 			className: this.props.itemClassName,
 			style: _style
 		};
 
-		return React.createElement("div",props,React.createElement(ItemRenderer,this.props.data));
-	}
-});
+		return (
+			<div {...props}>
+				<div>{this.props.data}</div>
+			</div>
+		);
 
-module.exports = Item;
+	}
+}
